Trait max7219::connectors::Connector

pub trait Connector {
    // Required methods
    fn devices(&self) -> usize;
    fn write_raw(
        &mut self,
        addr: usize,
        header: u8,
        data: u8
    ) -> Result<(), DataError>;

    // Provided method
    fn write_data(
        &mut self,
        addr: usize,
        command: Command,
        data: u8
    ) -> Result<(), DataError> { ... }
}

Describes the interface used to connect to the MX7219

Required Methods

fn devices(&self) -> usize

fn write_raw( &mut self, addr: usize, header: u8, data: u8 ) -> Result<(), DataError>

Writes data to given register as described by command

Arguments

addr - display to address as connected in series (0 -> last)

header - the command/register on the display to write to as u8

data - the data byte value to write

Errors

DataError - returned in case there was an error during data transfer

Provided Methods

fn write_data( &mut self, addr: usize, command: Command, data: u8 ) -> Result<(), DataError>

Writes data to given register as described by command

Arguments

addr - display to address as connected in series (0 -> last)

command - the command/register on the display to write to

data - the data byte value to write

Errors

DataError - returned in case there was an error during data transfer